Extract duplicated holiday-preset logic from DatePeriodPicker/Legacy into a shared util
