+++ src/ui/DatePeriodPickerLegacy.tsx
@@ -4,20 +4,22 @@ import {
   format,
   isSameDay,
   isSameMonth,
-  parse,
   startOfDay,
 } from 'date-fns';
 import de from 'date-fns/locale/de';
 import fr from 'date-fns/locale/fr';
 import nl from 'date-fns/locale/nl';
-import type { TFunction } from 'i18next';
 import { useMemo, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 
 import type { ApiHoliday } from '@/hooks/api/holidays';
 
-import { SupportedLanguages } from '../i18n';
 import type { Values } from '../types/Values';
+import {
+  computeHolidayPresets,
+  filterHolidaysForPreset,
+  parseHoliday,
+} from '../utils/holidayPresets';
 import { Button, ButtonVariants } from './Button';
 import { DatePicker } from './DatePicker';
 import type { InlineProps } from './Inline';
@@ -29,99 +31,7 @@ import { SwitchWithLabel } from './SwitchWithLabel';
 import { Text, TextVariants } from './Text';
 import { colors } from './theme';
 
-type HolidayPreset = {
-  label: string;
-  fetchStartDate: string;
-  fetchEndDate: string;
-  matchesHoliday: (holiday: {
-    type: string;
-    region?: string;
-    startDate: Date;
-  }) => boolean;
-};
-
 const locales = { nl, fr, de };
-
-const parseHoliday = (holiday: ApiHoliday, language: string, t: TFunction) => {
-  const name =
-    holiday.name[language as Values<typeof SupportedLanguages>] ?? '';
-  const regionLabel = holiday.region
-    ? t(`date_period_picker.region.${holiday.region}`)
-    : undefined;
-  return {
-    type: holiday.type,
-    region: holiday.region,
-    name: regionLabel ? `${name} (${regionLabel})` : name,
-    startDate: parse(holiday.startDate, 'yyyy-MM-dd', new Date()),
-    endDate: parse(holiday.endDate, 'yyyy-MM-dd', new Date()),
-  };
-};
-
-const getAcademicYearStart = (date: Date): number =>
-  date.getMonth() >= 7 ? date.getFullYear() : date.getFullYear() - 1;
-
-const formatAcademicYearLabel = (start: number) =>
-  `'${String(start).slice(-2)}-'${String(start + 1).slice(-2)}`;
-
-const publicHolidayPreset = (year: number, t: TFunction): HolidayPreset => ({
-  label: `${t('date_period_picker.quick_links.public_holidays')} ${year}`,
-  fetchStartDate: `${year}-01-01`,
-  fetchEndDate: `${year}-12-31`,
-  matchesHoliday: (holiday) =>
-    holiday.type !== 'schoolHolidays' &&
-    holiday.startDate.getFullYear() === year,
-});
-
-const schoolHolidayPreset = (
-  academicStart: number,
-  region: 'NL' | 'FR',
-  label: string,
-): HolidayPreset => ({
-  label: `${label} ${formatAcademicYearLabel(academicStart)}`,
-  fetchStartDate: `${academicStart}-08-01`,
-  fetchEndDate: `${academicStart + 1}-07-31`,
-  matchesHoliday: (holiday) =>
-    holiday.type === 'schoolHolidays' &&
-    holiday.region === region &&
-    getAcademicYearStart(holiday.startDate) === academicStart,
-});
-
-const filterHolidaysForPreset = (
-  holidays: ApiHoliday[],
-  preset: HolidayPreset,
-  language: string,
-  t: TFunction,
-) =>
-  holidays
-    .map((holiday) => parseHoliday(holiday, language, t))
-    .filter(
-      (holiday) =>
-        holiday.endDate >= new Date() && preset.matchesHoliday(holiday),
-    );
-
-const computeHolidayPresets = (today: Date, t: TFunction): HolidayPreset[] => {
-  const year = today.getFullYear();
-  const academicYear = getAcademicYearStart(today);
-  const schoolRegions = [
-    {
-      region: 'NL' as const,
-      label: t('date_period_picker.quick_links.flemish_school_holidays'),
-    },
-    {
-      region: 'FR' as const,
-      label: t('date_period_picker.quick_links.french_school_holidays'),
-    },
-  ];
-
-  return [
-    publicHolidayPreset(year, t),
-    publicHolidayPreset(year + 1, t),
-    ...schoolRegions.flatMap(({ region, label }) => [
-      schoolHolidayPreset(academicYear, region, label),
-      schoolHolidayPreset(academicYear + 1, region, label),
-    ]),
-  ];
-};
 
 type QuickLinkPeriod = { startDate: Date; endDate: Date; name: string };
